Update rebuild commands and routing in project.json and default.conf
The SSO project rebuild process is restructured in project.json including the 'server-sso:build' command and restarting nginx after building the app. This is done to ensure a correct build process with required dependencies, and to apply potential configuration changes immediately. Also, the linting configuration is removed as we are focusing on more significant issues currently. In terms of the routing defined in default.conf, there are modifications done in the order and addition of new routes offering a more logical and clean routing process for the app. These adjustments are essential for stable and effective app operation.