Remove dist/apps/io/chat and dist/apps/io/connect JavaScript files
Both files under the io directory namely, chat and connect were removed as part of the code cleanup process. This was done because they are no longer needed in the application. It’s good practice to remove deprecated or unused code to reduce confusion and maintainability in the future. This will save time for other developers or for yourself when you revisit the code in the future. This will also make your application smaller and faster.